added authenticate.html, authentication.py for further authentication process, updated requirements.txt, changed project structure
