+++ app.py
@@ -1,9 +1,9 @@
 import gc
-from flask import Flask, render_template, flash, request, url_for, redirect, session
-from wtforms import Form, validators, StringField, PasswordField, BooleanField
+from flask import Flask, render_template, flash, url_for, redirect, session
+
+from content.authentication import Authentication
+from content.authorization import Authorization
 from content_management import Content
-from dbconnect import connection
-from passlib.hash import sha256_crypt
 
 TOPIC_DICT = Content()
 
@@ -35,91 +35,16 @@ def dashboard():
     return render_template("content_dashboard.html", TOPIC_DICT=TOPIC_DICT)
 
 
-class RegistrationForm(Form):
-    username = StringField('Username', [validators.Length(min=4, max=20)])
-    email = StringField('Email Address', [validators.Length(min=6, max=20)])
-    password = PasswordField('Password', [validators.DataRequired(),
-                                          validators.EqualTo('confirm', message="Password must match.")])
-    confirm = PasswordField('Repeat Password')
-
-    accept_tos = BooleanField('I accept the Terms of Service and Privacy Notice (updated Jan 22, 2015)',
-                              [validators.DataRequired()])
-
-
 @app.route('/register/', methods=['GET', 'POST'])
-def register_page():
-    try:
-        form = RegistrationForm(request.form)
-
-        if request.method == "POST" and form.validate():
-            username = form.username.data
-            email = form.email.data
-
-            # TODO: sha512 registration maybe?
-            password = sha256_crypt.hash((str(form.password.data)))
-            c, conn = connection()
-
-            x = c.execute("SELECT * FROM users WHERE username = (%s);", (username,))
-            row_count = c.rowcount
-
-            if row_count == 0:
-                query = "INSERT INTO users (username, password, email, tracking) VALUES (%s, %s, %s, %s);"
-                values = (username, password, email, "/introduction-to-python-programming/")
-                c.execute(query, values)
-                conn.commit()
-                flash("Thanks for registering!")
-                c.close()
-                conn.close()
-                gc.collect()
-
-                session['logged_in'] = True
-                session['username'] = username
-
-                return redirect(url_for('dashboard'))
-
-            else:
-                flash("That username is already taken, please choose another!")
-                return render_template('register.html', form=form)
-
-        else:
-            return render_template('register.html', form=form)
-
-    except Exception as e:
-        return str(e)
+def register():
+    template = Authorization.register_page()
+    return template
 
 
 @app.route('/login/', methods=['GET', 'POST'])
-def login_page():
-    error = ''
-    try:
-        c, conn = connection()
-        if request.method == 'POST':
-            data = c.execute("SELECT * FROM users WHERE username = (%s);", (request.form['username'],))
-
-            if c.rowcount != 0:  # if user exists
-                data_password = c.fetchone()[2]  # get his password
-
-                # TODO: if sha512 registration, then change here also
-                if sha256_crypt.verify(request.form['password'], data_password):  # check his password
-                    session['logged_in'] = True
-                    session['username'] = request.form['username']
-
-                    flash("You are now logged in")
-                    return redirect(url_for("dashboard"))
-
-                else:
-                    error = "Invalid credentials, try again!"
-
-            else:
-                error = "Invalid credentials, try again!"
-
-        gc.collect()
-        return render_template('login.html', error=error)
-
-    except Exception as e:
-        flash(e)
-        error = "Invalid credentials, try again!"
-        return render_template("login.html", error=error)
+def login():
+    template = Authorization.login_page()
+    return template
 
 
 @app.route("/logout/")
@@ -131,17 +56,21 @@ def logout():
     return redirect(url_for('dashboard'))
 
 
+@app.route("/authenticate/", methods=['GET', 'POST'])
+def authentication():
+    template = Authentication.check_email()
+    return template
+
+
 # ONLY Error handling below #
-
-
 @app.errorhandler(404)
 def page_not_found(e):
-    return render_template("404.html")
+    return render_template("errors/404.html")
 
 
 @app.errorhandler(405)
 def method_not_found(e):
-    return render_template("405.html")
+    return render_template("errors/405.html")
 
 
 if __name__ == "__main__":
