+++ content/forms/registration_form.py
@@ -1,0 +1,12 @@
+from wtforms import Form, StringField, PasswordField, validators, BooleanField
+
+
+class RegistrationForm(Form):
+    username = StringField('Username', [validators.Length(min=4, max=20)])
+    email = StringField('Email Address', [validators.Length(min=6, max=20)])
+    password = PasswordField('Password', [validators.DataRequired(),
+                                          validators.EqualTo('confirm', message="Password must match.")])
+    confirm = PasswordField('Repeat Password')
+
+    accept_tos = BooleanField('I accept the Terms of Service and Privacy Notice (updated Jan 22, 2015)',
+                              [validators.DataRequired()])
