+++ content_management.py
@@ -1,11 +1,11 @@
 # TODO: This is nice idea and works pretty well, but I think we should:
 #  1. Cover this in class,
 #  2. Define similar dictionary,
-#  3. Get proper return functions
+#  3. Get proper return content
 
 def Content():
     TOPIC_DICT = {"Basics": [["Introduction to Python", "/introduction-to-python-programming/"],
-                             ["Print functions and Strings", "/python-tutorial-print-function-strings/"],
+                             ["Print content and Strings", "/python-tutorial-print-function-strings/"],
                              ["Math basics with Python 3", "/math-basics-python-3-beginner-tutorial/"]],
                   "Web Dev": []}
