+++ requirements.txt
@@ -7,3 +7,4 @@ passlib==1.7.2
 Werkzeug==1.0.1
 WTForms==2.3.1
 mysql-connector-python~=8.0.20
+requests~=2.23.0
