+++ templates/authenticate.html
@@ -1,0 +1,48 @@
+<!doctype html>
+<html lang="en">
+  <head>
+    <meta charset="utf-8">
+    <meta name="viewport" content="width=device-width, initial-scale=1, shrink-to-fit=no">
+    <meta name="description" content="">
+    <meta name="author" content="Mark Otto, Jacob Thornton, and Bootstrap contributors">
+    <meta name="generator" content="Jekyll v4.0.1">
+    <title>Authenticate - V-Biometrics</title>
+
+    <link rel="canonical" href="https://getbootstrap.com/docs/4.5/examples/sign-in/">
+
+    <!-- Bootstrap core CSS -->
+<link href="{{ url_for('static', filename='css/bootstrap.css') }}" rel="stylesheet">
+
+    <style>
+      .bd-placeholder-img {
+        font-size: 1.125rem;
+        text-anchor: middle;
+        -webkit-user-select: none;
+        -moz-user-select: none;
+        -ms-user-select: none;
+        user-select: none;
+      }
+
+      @media (min-width: 768px) {
+        .bd-placeholder-img-lg {
+          font-size: 3.5rem;
+        }
+      }
+    </style>
+    <!-- Custom styles for this template -->
+    <link href="{{ url_for('static', filename='css/signin.css') }}" rel="stylesheet">
+  </head>
+  <body class="text-center" style="background-color:#f0f0f0;">
+    <form class="form-signin" method="post">
+      <img class="img-fluid" src="{{ url_for('static', filename='images/logos/INV_logo_alone_2.png') }}" alt="" width="120" height="120">
+      <h1 class="h3 mb-3 font-weight-normal">Enter your email:</h1>
+      <label for="inputEmail" class="sr-only">Email test</label>
+      <input type="email" id="inputEmail" class="form-control" placeholder="Email address" name="email" value="{{request.form.email}}" required autofocus>
+      <button class="btn btn-lg btn-warning btn-block" type="submit">Continue</button>
+      <p class="text-dark bg-danger font-weight-bold">{{error}}</p>
+      <a href="/register/" class="text-dark">No account? Sign up here!</a>
+      <p class="mt-5 mb-3 text-muted">&copy; 2019-2020 Malinowski, Wieczorek</p>
+      <a class="btn btn-primary" href="/" role="button">Return to Homepage</a>
+    </form>
+  </body>
+</html>
