+++ templates/main.html
@@ -2,13 +2,13 @@
 
 {% block body %}
 
-<body style="background-color:black;">
+<body style="background-color:white;">
     <script src="//code.jquery.com/jquery-3.5.1.min.js"></script>
     <script type="text/javascript" src="{{ url_for('static', filename='js/bootstrap.min.js') }}"></script>
 
     <div class="container" align="center">
         <a href="/dashboard/">
-            <img src="{{ url_for('static', filename='images/logos/logo_full_on_black_2.png') }}" style="height:600px;">
+            <img src="{{ url_for('static', filename='images/logos/INV_logo_full_on_black_2.png') }}" style="height:600px;">
         </a>
     </div>
